Add custom route for getting dropped file
